Move PROXY proto logic to its own `proxyproto.go` file.
